python_client/requirements.txt
Flask==1.0.2
web3==4.5.0
py-solc==3.1.0
flask-redis==0.3.0